requirements.txt
allure-pytest==2.13.5
jsonschema==4.23.0
jsonref==1.1.0
pytest==8.3.3
pytest-html==4.1.1
pyyaml==6.0.2
requests==2.32.3